+++ src/pro/blocks/index.js
@@ -5,4 +5,3 @@ import './add-to-cart-button/index.js';
 import './business-hours/index.js';
 import './review-comparison/index.js';
 import './woo-comparison/index.js';
-import './countdown/index';
