+++ src/pro/plugins/index.js
@@ -8,3 +8,4 @@ import './posts/index.js';
 import './popup/index.js';
 import './sticky/index.js';
 import './wc-integration/index.js';
+import './countdown/index';
